+++ dev-docs/engineering/event-catalog.md
@@ -6,7 +6,7 @@
 | --- | --- |
 | `submission` | noj-core/src/mq/consumer.ts<br>noj-core/src/routes/sse.ts |
 | `queue` | noj-core/src/mq/consumer.ts<br>noj-core/src/routes/sse.ts<br>noj-core/src/services/queue.ts<br>noj-core/src/services/self-tests.ts<br>noj-core/src/services/submissions/artifact-submissions.ts<br>noj-core/src/services/submissions/submissions-crud.ts<br>noj-core/src/services/submissions/submissions-rejudge.ts |
-| `user` | noj-core/src/routes/conversations.ts<br>noj-core/src/routes/sse.ts<br>noj-core/src/services/messages.ts<br>noj-core/src/services/notifications.ts |
+| `user` | noj-core/src/domains/messaging/routes/conversations.ts<br>noj-core/src/domains/messaging/services/messages.ts<br>noj-core/src/routes/sse.ts<br>noj-core/src/services/notifications.ts |
 | `contestRanking` | noj-core/src/routes/sse.ts<br>noj-core/src/services/submissions/submissions-result.ts |
 | `contestSubmission` | noj-core/src/routes/sse.ts<br>noj-core/src/services/submissions/artifact-submissions.ts<br>noj-core/src/services/submissions/submissions-crud.ts |
 | `stats` | noj-core/src/routes/sse.ts<br>noj-core/src/services/stats-cache.ts |
