+++ dev-docs/engineering/route-catalog.md
@@ -9,7 +9,6 @@
 | DELETE | `/:id` | noj-core/src/routes/problems.ts |
 | DELETE | `/:id` | noj-core/src/routes/tags.ts |
 | DELETE | `/:id` | noj-core/src/routes/trainings.ts |
-| DELETE | `/:id/messages/:messageId` | noj-core/src/routes/conversations.ts |
 | DELETE | `/:id/problems/:problemId` | noj-core/src/routes/trainings.ts |
 | DELETE | `/:id/questions/:qid` | noj-core/src/routes/problems.ts |
 | DELETE | `/:id/support-package` | noj-core/src/routes/problems.ts |
@@ -23,7 +22,6 @@
 | GET | `/` | noj-core/src/routes/admin-announcements.ts |
 | GET | `/` | noj-core/src/routes/admin-trainings.ts |
 | GET | `/` | noj-core/src/routes/announcements.ts |
-| GET | `/` | noj-core/src/routes/conversations.ts |
 | GET | `/` | noj-core/src/routes/problems.ts |
 | GET | `/` | noj-core/src/routes/queue.ts |
 | GET | `/` | noj-core/src/routes/rankings.ts |
@@ -36,14 +34,12 @@
 | GET | `/:id` | noj-core/src/routes/submissions.ts |
 | GET | `/:id` | noj-core/src/routes/trainings.ts |
 | GET | `/:id/avatar` | noj-core/src/routes/users.ts |
-| GET | `/:id/messages` | noj-core/src/routes/conversations.ts |
 | GET | `/:id/problems` | noj-core/src/routes/trainings.ts |
 | GET | `/:id/profile` | noj-core/src/routes/users.ts |
 | GET | `/:id/questions` | noj-core/src/routes/problems.ts |
 | GET | `/:id/status` | noj-core/src/routes/submissions.ts |
 | GET | `/:id/support-package` | noj-core/src/routes/problems.ts |
 | GET | `/:id/template` | noj-core/src/routes/problems.ts |
-| GET | `/:id/unread-count` | noj-core/src/routes/conversations.ts |
 | GET | `/admin/boards/:boardId/role-grants` | noj-core/src/routes/community-admin.ts |
 | GET | `/admin/comments/pending` | noj-core/src/routes/community-admin.ts |
 | GET | `/admin/reports` | noj-core/src/routes/community-admin.ts |
@@ -58,7 +54,6 @@
 | GET | `/containing` | noj-core/src/routes/trainings.ts |
 | GET | `/contests/:id/events` | noj-core/src/routes/sse.ts |
 | GET | `/events` | noj-core/src/routes/announcements.ts |
-| GET | `/events` | noj-core/src/routes/conversations.ts |
 | GET | `/feed` | noj-core/src/routes/community.ts |
 | GET | `/health` | noj-core/src/routes/health.ts |
 | GET | `/history` | noj-core/src/routes/checkin.ts |
@@ -91,7 +86,6 @@
 | GET | `/today` | noj-core/src/routes/checkin.ts |
 | GET | `/today-stats` | noj-core/src/routes/submissions.ts |
 | GET | `/total-stats` | noj-core/src/routes/submissions.ts |
-| GET | `/unread-count` | noj-core/src/routes/conversations.ts |
 | GET | `jti` | noj-core/src/routes/auth.ts |
 | GET | `jti` | noj-core/src/routes/auth.ts |
 | GET | `NOJ_ENV` | noj-core/src/routes/auth.ts |
@@ -122,15 +116,6 @@
 | GET | `userId` | noj-core/src/routes/community.ts |
 | GET | `userId` | noj-core/src/routes/community.ts |
 | GET | `userId` | noj-core/src/routes/community.ts |
-| GET | `userId` | noj-core/src/routes/conversations.ts |
-| GET | `userId` | noj-core/src/routes/conversations.ts |
-| GET | `userId` | noj-core/src/routes/conversations.ts |
-| GET | `userId` | noj-core/src/routes/conversations.ts |
-| GET | `userId` | noj-core/src/routes/conversations.ts |
-| GET | `userId` | noj-core/src/routes/conversations.ts |
-| GET | `userId` | noj-core/src/routes/conversations.ts |
-| GET | `userId` | noj-core/src/routes/conversations.ts |
-| GET | `userId` | noj-core/src/routes/conversations.ts |
 | GET | `userId` | noj-core/src/routes/problems.ts |
 | GET | `userId` | noj-core/src/routes/problems.ts |
 | GET | `userId` | noj-core/src/routes/problems.ts |
@@ -193,16 +178,13 @@
 | PATCH | `/posts/:postId` | noj-core/src/routes/community.ts |
 | POST | `/` | noj-core/src/routes/admin-announcements.ts |
 | POST | `/` | noj-core/src/routes/checkin.ts |
-| POST | `/` | noj-core/src/routes/conversations.ts |
 | POST | `/` | noj-core/src/routes/problems.ts |
 | POST | `/` | noj-core/src/routes/submissions.ts |
 | POST | `/` | noj-core/src/routes/tags.ts |
 | POST | `/` | noj-core/src/routes/trainings.ts |
 | POST | `/:id/merge` | noj-core/src/routes/tags.ts |
-| POST | `/:id/messages` | noj-core/src/routes/conversations.ts |
 | POST | `/:id/problems` | noj-core/src/routes/trainings.ts |
 | POST | `/:id/questions` | noj-core/src/routes/problems.ts |
-| POST | `/:id/read` | noj-core/src/routes/conversations.ts |
 | POST | `/:id/submit` | noj-core/src/routes/problems.ts |
 | POST | `/admin/boards` | noj-core/src/routes/community-admin.ts |
 | POST | `/admin/comments/:commentId/:status` | noj-core/src/routes/community-admin.ts |
