+++ noj-core/src/domains/community/index.ts
@@ -1,2 +1,5 @@
-export { createActivity } from "../../services/community/community.ts";
+export {
+  createActivity,
+  getCommunityConfig,
+} from "../../services/community/community.ts";
 export { createNotification } from "../../services/notifications.ts";
